DEV: Slit calculator - add in disable slit scaling
If a default is set to disable the scaling of certain slits - i.e. say S3 is fixed to stop over illumination of a mirror - then it will only scale the slits that do not have this flag enabled.

Additionally, add the global default for which slit pair will be used as the collimation slits (when calculating the slit openings from a defined footprint and resolution)